pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weekly-slideshow-engine"
version = "0.0.0"
description = "Weekly slideshow engine (MVP skeleton)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Your Name"}]
keywords = ["video", "cli"]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
test = ["pytest"]
image = ["Pillow", "pillow-heif"]
render = ["moviepy"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "render: marks tests that require rendering (video encoding)",
]
testpaths = ["tests"]
addopts = "-v --tb=short"
